requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
joblib
streamlit
apache-airflow
psycopg2-binary
sqlalchemy